Avoid the creation of a notnull constraint
I don't know if it's the auto-validation added since symfony4.3 that does that. 
but if I don't add the ? in the annotation: 
@var File   ->  @var ?File
I have a notnull constraint that prevents the normal behaviour of the form even if it is configured with required=false
When I edit an uploadable without wanting to change the image, the validator blocks on "imageFile can't be null".